server/requirements.txt
fastapi
uvicorn[standard]
requests
google-genai
python-dotenv
python-jose[cryptography]
bcrypt
python-multipart
pymongo